requirements.txt
Flask==1.1.1
howdoi==1.2.1
python-telegram-bot==12.2.0

